chore: move skill schemas to skill_schemas file and make other minor update
