+++ src/repositories/__init__.py
@@ -1,2 +1,3 @@
 from .general_user_repo import UserRepository
+from .skill_repo import SkillRepository
 from .base_repo import BaseRepository
